// src/types/models/Body_create_image_api_images_create_img_product_post.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Body_create_image_api_images_create_img_product_post = {
    description: string;
    image_file: Blob;
    product_id: number;
    main_image: boolean;
};